4.21

+++ part4/bloglist/controllers/blogs.js
@@ -55,13 +55,22 @@ blogsRouter.post("/", async (request, response, next) => {
 //delete blog, refactored
 blogsRouter.delete("/:id", async (request, response, next) => {
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log(request)
+  const user = await User.findById(decodedToken.id);
+  const userId = await user.id;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({error: "wrong user"})
+  }
   // if (!request.token || !decodedToken.id) {
   //   response.status(401).json({ error: "token missing or invalid" });
   // } else if (blog.user.toString() !== userid.toString()) {
   //   response.status(401).json({ error: "wrong user" });
   // }
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
